+++ example/data_2.js
@@ -1,0 +1,35 @@
+const geojson2 = [{
+    "type": "FeatureCollection",
+    "features": [{
+        "type": "Feature",
+        "geometry": {
+            "type": "LineString",
+            "coordinates": [
+                [8.6865264, 49.3859188, 114.5],
+                [8.6864108, 49.3868472, 114.3],
+                [8.6860538, 49.3903808, 114.8]
+            ]
+        },
+        "properties": {
+            "attributeType": "3"
+        }
+    }, {
+        "type": "Feature",
+        "geometry": {
+            "type": "LineString",
+            "coordinates": [
+                [8.6860538, 49.3903808, 114.8],
+                [8.6857921, 49.3936309, 114.4],
+                [8.6860124, 49.3936431, 114.3]
+            ]
+        },
+        "properties": {
+            "attributeType": "0"
+        }
+    }],
+    "properties": {
+        "Creator": "OpenRouteService.org",
+        "records": 2,
+        "summary": "steepness"
+    }
+}]